Move the CD ripper into its own directory.
