+++ src/CMakeLists.txt
@@ -1035,14 +1035,14 @@ optional_source(HAVE_AUDIOCD
     devices/cddadevice.cpp
     devices/cddalister.cpp
     devices/cddasongloader.cpp
-    ui/ripcd.cpp
+    ripper/ripcd.cpp
   HEADERS
     devices/cddadevice.h
     devices/cddalister.h
     devices/cddasongloader.h
-    ui/ripcd.h
+    ripper/ripcd.h
   UI
-    ui/ripcd.ui
+    ripper/ripcd.ui
 )
 
 # mtp device
