+++ src/ripper/ripcd.cpp
