+++ src/ripper/ripcd.h
