+++ src/ripper/ripcd.ui
